Initial pass at Fluent integration
Integrate Fluent into Pokétwo. Because I'm insisting on externalizing
the command formatting (for flexibility, future-proofing and
maintainability reasons), we have to do a small hack to pass the clean
prefix through to Fluent---but I think it's OK.

I'm also deciding to swallow localization errors and log them because
I'd rather the message ID make it to a user's eyes than the command
break entirely.