Resolve group, cog, and command help in Fluent
Add the missing functionality to resolve command/group help and cog
descriptions in Fluent. We fall back to the docstring if it's missing in
Fluent, but make sure to log a warning.

We're using a `.help`/`.description` attribute to future-proof against
potentially localizing more than the help messages/descriptions. Other
attributes can be added to the same logical message if so, preserving
grouping and organization.