Unignore cogs/config.py
It seems unintentional, but since "config.py" is ignored by Git, so is
"cogs/config.py", which is an actual cog that the bot loads. Add a line
to explicitly allow "cogs/config.py".

This is also useful in general because a lot of tools automatically
ignore things that are additionally ignored by Git (such as ripgrep).