Add support for defining skeleton embeds in Fluent
Add a system to define the skeleton of Discord embeds inside of Fluent.
This is somewhat of an involved process that requires understanding of
some edge cases, so I made sure to author a thorough docstring for
people reading the code in the future.

Simple embeds may be defined entirely within Fluent, but they are often
mutated after the fact to add dynamic fields and other attributes. There
are facilities provided in the method to strike a balance between
defining some parts in Fluent while cooperating with dynamic values that
can only be accessed from code.